auth_service/requirements.txt
bcrypt==4.0.1
blinker==1.9.0
click==8.1.7
dnspython==2.7.0
Flask==2.3.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
marshmallow==3.23.1
packaging==24.2
PyJWT==2.8.0
pymongo==4.5.0
python-dotenv==1.0.1
Werkzeug==3.1.3
